Tune Vim to be just a simple text editor
I have decided that I don't want powerful batteries to be included. Vim
can just suit me as a text editor for performing quick changes to files.
It doesn't need to function like an IDE, like providing code completion.
Actually, it would be wrong to turn it into an IDE; VimScript is far too
weak of a scripting language to be capable of doing that. Emacs Lisp and
its heavy integration to Emacs on the other hand, is mightier and more
competent for the job.