Switch to termite as the default terminal emulator
The reason is small memory consumption in contrast to the previous emulator.
Alacritty, a GPU-accelerated terminal emulator, at a minimum uses 60MB-100MB of memory. Sometimes even more.
I find this ridiculous. Sure, the emulator is fast, but I believe that memory could be utilized for other programs already running, or that will be run.
In my tests, termite uses around 25-30MB. Memory usage that is similar/close to the flagship terminal emulator of Xorg, xterm, and other popular emulators, such as urxvt and st.
It also isn't a pain to configure like the aforementioned emulators.